refactor(file): Move descriptors to its own file.
